+++ src/main/java/com/github/icecheesecat/kantaicraft/entityship/stance/HostileStance.java
@@ -12,7 +12,7 @@ public interface HostileStance extends Stance {
     @Override
     default void setupSyncedDataFromStance(SynchedEntityData entityData, RandomSource random) {
         entityData.define(EntityShip.DATA_AIRCRAFT, Integer.MAX_VALUE);
-        entityData.define(EntityShip.DATA_AMMO, Float.MAX_VALUE);
+        entityData.define(EntityShip.DATA_AMMO, Integer.MAX_VALUE);
         entityData.define(EntityShip.DATA_SHOULD_PICK_UP_ITEM, false);
     }
